backend/requirements.txt
fastapi
uvicorn
openai
pinecone-client
redis
anthropic
langchain
tiktoken
